+++ src/sdk/conference/channel.js
@@ -21,6 +21,7 @@ import {Subscription} from './subscription.js';
 import {ConferenceError} from './error.js';
 import * as Utils from '../base/utils.js';
 import * as SdpUtils from '../base/sdputils.js';
+import {TransportSettings, TransportType} from '../base/transport.js';
 
 /**
  * @class ConferencePeerConnectionChannel
@@ -875,12 +876,19 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
     const internalId = this._reverseIdMap.get(sessionId);
     if (this._subscribePromises.has(internalId)) {
       const mediaStream = this._remoteMediaStreams.get(sessionId);
-      const subscription = new Subscription(sessionId, mediaStream, () => {
-        this._unsubscribe(internalId);
-      }, () => this._getStats(),
-      (trackKind) => this._muteOrUnmute(sessionId, true, false, trackKind),
-      (trackKind) => this._muteOrUnmute(sessionId, false, false, trackKind),
-      (options) => this._applyOptions(sessionId, options));
+      const transportSettings =
+          new TransportSettings(TransportType.WEBRTC, this._id);
+      transportSettings.rtpTransceivers =
+          this._subscribeTransceivers.get(internalId).transceivers;
+      const subscription = new Subscription(
+          sessionId, mediaStream, transportSettings,
+          () => {
+            this._unsubscribe(internalId);
+          },
+          () => this._getStats(),
+          (trackKind) => this._muteOrUnmute(sessionId, true, false, trackKind),
+          (trackKind) => this._muteOrUnmute(sessionId, false, false, trackKind),
+          (options) => this._applyOptions(sessionId, options));
       this._subscriptions.set(sessionId, subscription);
       // Resolve subscription if mediaStream is ready.
       if (this._subscriptions.get(sessionId).stream) {
@@ -888,12 +896,20 @@ export class ConferencePeerConnectionChannel extends EventDispatcher {
         this._subscribePromises.delete(internalId);
       }
     } else if (this._publishPromises.has(internalId)) {
-      const publication = new Publication(sessionId, () => {
-        this._unpublish(internalId);
-        return Promise.resolve();
-      }, () => this._getStats(),
-      (trackKind) => this._muteOrUnmute(sessionId, true, true, trackKind),
-      (trackKind) => this._muteOrUnmute(sessionId, false, true, trackKind));
+      const transportSettings =
+          new TransportSettings(TransportType.WEBRTC, this._id);
+      transportSettings.transceivers =
+          this._publishTransceivers.get(internalId).transceivers;
+      const publication = new Publication(
+          sessionId,
+          transportSettings,
+          () => {
+            this._unpublish(internalId);
+            return Promise.resolve();
+          },
+          () => this._getStats(),
+          (trackKind) => this._muteOrUnmute(sessionId, true, true, trackKind),
+          (trackKind) => this._muteOrUnmute(sessionId, false, true, trackKind));
       this._publications.set(sessionId, publication);
       this._publishPromises.get(internalId).resolve(publication);
       // Do not fire publication's ended event when associated stream is ended.