+++ src/sdk/conference/subscription.js
@@ -270,7 +270,8 @@ export class SubscriptionUpdateOptions {
  */
 export class Subscription extends EventDispatcher {
   // eslint-disable-next-line require-jsdoc
-  constructor(id, stream, stop, getStats, mute, unmute, applyOptions) {
+  constructor(
+      id, stream, transport, stop, getStats, mute, unmute, applyOptions) {
     super();
     if (!id) {
       throw new TypeError('ID cannot be null or undefined.');
@@ -297,6 +298,18 @@ export class Subscription extends EventDispatcher {
       // MediaStream may not be available.
       writable: true,
       value: stream,
+    });
+    /**
+     * @member {Owt.Base.TransportSettings} transport
+     * @instance
+     * @readonly
+     * @desc Transport settings for the subscription.
+     * @memberof Owt.Base.Subscription
+     */
+    Object.defineProperty(this, 'transport', {
+      configurable: false,
+      writable: false,
+      value: transport,
     });
     /**
      * @function stop